SNO-31 Add helper.py and remove helpers from search.py
