+++ src/snowflakes/search.py
@@ -1,5 +1,9 @@
 import re
+from urllib.parse import urlencode
 from pyramid.view import view_config
+from pyramid.httpexceptions import HTTPBadRequest
+from pyramid.security import effective_principals
+from elasticsearch.helpers import scan
 from snovault import (
     AbstractCollection,
     TYPES,
@@ -7,11 +11,23 @@ from snovault import (
 from snovault.elasticsearch import ELASTIC_SEARCH
 from snovault.elasticsearch.interfaces import RESOURCES_INDEX
 from snovault.resource_views import collection_view_listing_db
-from elasticsearch.helpers import scan
-from pyramid.httpexceptions import HTTPBadRequest
-from pyramid.security import effective_principals
-from urllib.parse import urlencode
-from collections import OrderedDict
+from snovault.helpers.helper import (
+    sort_query,
+    get_pagination,
+    get_filtered_query,
+    prepare_search_term,
+    set_sort_order,
+    get_search_fields,
+    list_visible_columns_for_schemas,
+    list_result_fields,
+    set_filters,
+    set_facets,
+    format_results,
+    search_result_actions,
+    format_facets,
+    normalize_query,
+    iter_long_json
+)
 
 
 CHAR_COUNT = 32
@@ -41,552 +57,6 @@ DEFAULT_DOC_TYPES = [
     'Snowflake',
 ]
 
-
-def get_pagination(request):
-    from_ = request.params.get('from') or 0
-    size = request.params.get('limit', 25)
-    if size in ('all', ''):
-        size = None
-    else:
-        try:
-            size = int(size)
-        except ValueError:
-            size = 25
-    return from_, size
-
-
-def get_filtered_query(term, search_fields, result_fields, principals, doc_types):
-    return {
-        'query': {
-            'query_string': {
-                'query': term,
-                'fields': search_fields,
-                'default_operator': 'AND'
-            }
-        },
-        'post_filter': {
-            'bool': {
-                'must': [
-                    {
-                        'terms': {
-                            'principals_allowed.view': principals
-                        }
-                    },
-                    {
-                        'terms': {
-                            'embedded.@type': doc_types
-                        }
-                    }
-                ],
-                'must_not': []
-            }
-        },
-
-        '_source': list(result_fields),
-    }
-
-
-def prepare_search_term(request):
-    from antlr4 import IllegalStateException
-    from lucenequery.prefixfields import prefixfields
-    from lucenequery import dialects
-
-    search_term = request.params.get('searchTerm', '').strip() or '*'
-    if search_term == '*':
-        return search_term
-
-    # avoid interpreting slashes as regular expressions
-    search_term = search_term.replace('/', r'\/')
-    # elasticsearch uses : as field delimiter, but we use it as namespace designator
-    # if you need to search fields you have to use @type:field
-    # if you need to search fields where the field contains ":", you will have to escape it
-    # yourself
-    if search_term.find("@type") < 0:
-        search_term = search_term.replace(':', '\:')
-    try:
-        query = prefixfields('embedded.', search_term, dialects.elasticsearch)
-    except (IllegalStateException):
-        msg = "Invalid query: {}".format(search_term)
-        raise HTTPBadRequest(explanation=msg)
-    else:
-        return query.getText()
-
-
-def set_sort_order(request, search_term, types, doc_types, query, result):
-    """
-    sets sort order for elasticsearch results
-    """
-    sort = OrderedDict()
-    result_sort = OrderedDict()
-
-    # Prefer sort order specified in request, if any
-    requested_sort = request.params.get('sort')
-    if requested_sort:
-        if requested_sort.startswith('-'):
-            name = requested_sort[1:]
-            order = 'desc'
-        else:
-            name = requested_sort
-            order = 'asc'
-        # TODO: unmapped type needs to be determined, not hard coded
-        sort['embedded.' + name] = result_sort[name] = {
-            'order': order,
-            'unmapped_type': 'keyword',
-        }
-
-    # Otherwise we use a default sort only when there's no text search to be ranked
-    if not sort and search_term == '*':
-
-        # If searching for a single type, look for sort options in its schema
-        if len(doc_types) == 1:
-            type_schema = types[doc_types[0]].schema
-            if 'sort_by' in type_schema:
-                for k, v in type_schema['sort_by'].items():
-                    # Should always sort on raw field rather than analyzed field
-                    sort['embedded.' + k] = result_sort[k] = v
-
-        # Default is most recent first, then alphabetical by label
-        if not sort:
-            sort['embedded.date_created'] = result_sort['date_created'] = {
-                'order': 'desc',
-                'unmapped_type': 'keyword',
-            }
-            sort['embedded.label'] = result_sort['label'] = {
-                'order': 'asc',
-                'missing': '_last',
-                'unmapped_type': 'keyword',
-            }
-
-    if sort:
-        query['sort'] = sort
-        result['sort'] = result_sort
-        return True
-
-    return False
-
-
-def get_search_fields(request, doc_types):
-    """
-    Returns set of columns that are being searched and highlights
-    """
-    fields = {'uuid'}
-    highlights = {}
-    types = request.registry[TYPES]
-    for doc_type in doc_types:
-        type_info = types[doc_type]
-        for value in type_info.schema.get('boost_values', ()):
-            fields.add('embedded.' + value)
-            highlights['embedded.' + value] = {}
-    return list(fields), highlights
-
-
-def list_visible_columns_for_schemas(request, schemas):
-    """
-    Returns mapping of default columns for a set of schemas.
-    """
-    columns = OrderedDict({'@id': {'title': 'ID'}})
-    for schema in schemas:
-        if 'columns' in schema:
-            columns.update(schema['columns'])
-        else:
-            # default columns if not explicitly specified
-            columns.update(OrderedDict(
-                (name, {
-                    'title': schema['properties'][name].get('title', name)
-                })
-                for name in [
-                    '@id', 'title', 'description', 'name', 'accession',
-                    'aliases'
-                ] if name in schema['properties']
-            ))
-
-    fields_requested = request.params.getall('field')
-    if fields_requested:
-        limited_columns = OrderedDict()
-        for field in fields_requested:
-            if field in columns:
-                limited_columns[field] = columns[field]
-            else:
-                # We don't currently traverse to other schemas for embedded
-                # objects to find property titles. In this case we'll just
-                # show the field's dotted path for now.
-                limited_columns[field] = {'title': field}
-                for schema in schemas:
-                    if field in schema['properties']:
-                        limited_columns[field] = {
-                            'title': schema['properties'][field]['title']
-                        }
-                        break
-        columns = limited_columns
-
-    return columns
-
-
-def list_result_fields(request, doc_types):
-    """
-    Returns set of fields that are requested by user or default fields
-    """
-    frame = request.params.get('frame')
-    fields_requested = request.params.getall('field')
-    if fields_requested:
-        fields = {'embedded.@id', 'embedded.@type'}
-        fields.update('embedded.' + field for field in fields_requested)
-    elif frame in ['embedded', 'object']:
-        fields = [frame + '.*']
-    else:
-        frame = 'columns'
-        fields = {'embedded.@id', 'embedded.@type'}
-        if request.has_permission('search_audit'):
-            fields.add('audit.*')
-        types = request.registry[TYPES]
-        schemas = [types[doc_type].schema for doc_type in doc_types]
-        columns = list_visible_columns_for_schemas(request, schemas)
-        fields.update('embedded.' + column for column in columns)
-
-    # Ensure that 'audit' field is requested with _source in the ES query
-    if request.__parent__ and '/metadata/' in request.__parent__.url and request.has_permission('search_audit'):
-        fields.add('audit.*')
-
-    return fields
-
-
-def build_terms_filter(query_filters, field, terms, query):
-    if field.endswith('!'):
-        field = field[:-1]
-        if not field.startswith('audit'):
-            field = 'embedded.' + field
-        # Setting not filter instead of terms filter
-        if terms == ['*']:
-            negative_filter_condition = {
-                'exists': {
-                    'field': field,
-                }
-            }
-        else:
-            negative_filter_condition = {
-                'terms': {
-                    field: terms
-                }
-            }
-        query_filters['must_not'].append(negative_filter_condition)
-    else:
-        if not field.startswith('audit'):
-            field = 'embedded.' + field
-        if terms == ['*']:
-            filter_condition = {
-                'exists': {
-                    'field': field,
-                }
-            }
-        else:
-            filter_condition = {
-                'terms': {
-                    field: terms,
-                },
-            }
-        query_filters['must'].append(filter_condition)
-
-
-def set_filters(request, query, result, static_items=None):
-    """
-    Sets filters in the query
-    """
-    query_filters = query['post_filter']['bool']
-    used_filters = {}
-    if static_items is None:
-        static_items = []
-
-    # Get query string items plus any static items, then extract all the fields
-    qs_items = list(request.params.items())
-    total_items = qs_items + static_items
-    qs_fields = [item[0] for item in qs_items]
-    fields = [item[0] for item in total_items]
-
-    # Now make lists of terms indexed by field
-    all_terms = {}
-    for item in total_items:
-        if item[0] in all_terms:
-            all_terms[item[0]].append(item[1])
-        else:
-            all_terms[item[0]] = [item[1]]
-
-    for field in fields:
-        if field in used_filters:
-            continue
-
-        terms = all_terms[field]
-        if field in ['type', 'limit', 'y.limit', 'x.limit', 'mode', 'annotation',
-                     'format', 'frame', 'datastore', 'field', 'region', 'genome',
-                     'sort', 'from', 'referrer']:
-            continue
-
-        # Add filter to result
-        if field in qs_fields:
-            for term in terms:
-                qs = urlencode([
-                    (k.encode('utf-8'), v.encode('utf-8'))
-                    for k, v in qs_items
-                    if '{}={}'.format(k, v) != '{}={}'.format(field, term)
-                ])
-                result['filters'].append({
-                    'field': field,
-                    'term': term,
-                    'remove': '{}?{}'.format(request.path, qs)
-                })
-
-        if field == 'searchTerm':
-            continue
-
-        # Add to list of active filters
-        used_filters[field] = terms
-
-        # Add filter to query
-        build_terms_filter(query_filters, field, terms, query)
-
-    return used_filters
-
-
-def build_aggregation(facet_name, facet_options, min_doc_count=0):
-    """Specify an elasticsearch aggregation from schema facet configuration.
-    """
-    exclude = []
-    if facet_name == 'type':
-        field = 'embedded.@type'
-        exclude = ['Item']
-    elif facet_name.startswith('audit'):
-        field = facet_name
-    else:
-        field = 'embedded.' + facet_name
-    agg_name = facet_name.replace('.', '-')
-
-    facet_type = facet_options.get('type', 'terms')
-    if facet_type == 'terms':
-        agg = {
-            'terms': {
-                'field': field,
-                'min_doc_count': min_doc_count,
-                'size': 100,
-            },
-        }
-        if exclude:
-            agg['terms']['exclude'] = exclude
-    elif facet_type == 'exists':
-        agg = {
-            'filters': {
-                'filters': {
-                    'yes': {
-                        'bool': {
-                            'must': {
-                                'exists': {'field': field}
-                            }
-                        }
-                    },
-                    'no': {
-                        'bool': {
-                            'must_not': {
-                                'exists': {'field': field}
-                            }
-                        }
-                    },
-                },
-            },
-        }
-    else:
-        raise ValueError('Unrecognized facet type {} for {} facet'.format(
-            facet_type, field))
-
-    return agg_name, agg
-
-
-def set_facets(facets, used_filters, principals, doc_types):
-    """
-    Sets facets in the query using filters
-    """
-    aggs = {}
-    for facet_name, facet_options in facets:
-        # Filter facet results to only include
-        # objects of the specified type(s) that the user can see
-        filters = [
-            {'terms': {'principals_allowed.view': principals}},
-            {'terms': {'embedded.@type': doc_types}},
-        ]
-        negative_filters = []
-        # Also apply any filters NOT from the same field as the facet
-        for field, terms in used_filters.items():
-            if field.endswith('!'):
-                query_field = field[:-1]
-            else:
-                query_field = field
-
-            # if an option was selected in this facet,
-            # don't filter the facet to only include that option
-            if query_field == facet_name:
-                continue
-
-            if not query_field.startswith('audit'):
-                query_field = 'embedded.' + query_field
-
-            if field.endswith('!'):
-                if terms == ['*']:
-                    negative_filters.append({'exists': {'field': query_field}})
-                else:
-                    negative_filters.append({'terms': {query_field: terms}})
-            else:
-                if terms == ['*']:
-                    filters.append({'exists': {'field': query_field}})
-                else:
-                    filters.append({'terms': {query_field: terms}})
-
-        agg_name, agg = build_aggregation(facet_name, facet_options)
-        aggs[agg_name] = {
-            'aggs': {
-                agg_name: agg
-            },
-            'filter': {
-                'bool': {
-                    'must': filters,
-                    'must_not': negative_filters
-                },
-            },
-        }
-
-    return aggs
-
-
-def format_results(request, hits, result=None):
-    """
-    Loads results to pass onto UI
-    """
-    fields_requested = request.params.getall('field')
-    if fields_requested:
-        frame = 'embedded'
-    else:
-        frame = request.params.get('frame')
-
-    # Request originating from metadata generation will skip to
-    # partion of the code that adds audit  object to result items
-    if request.__parent__ and '/metadata/' in request.__parent__.url:
-        frame = ''
-
-    any_released = False  # While formatting, figure out if any are released.
-
-    if frame in ['embedded', 'object']:
-        for hit in hits:
-            if not any_released and hit['_source'][frame].get('status','released') == 'released':
-                any_released = True
-            yield hit['_source'][frame]
-    else:
-        # columns
-        for hit in hits:
-            item = hit['_source']['embedded']
-            if not any_released and item.get('status','released') == 'released':
-                any_released = True # Not exp? 'released' to do the least harm
-            if 'audit' in hit['_source']:
-                item['audit'] = hit['_source']['audit']
-            if 'highlight' in hit:
-                item['highlight'] = {}
-                for key in hit['highlight']:
-                    item['highlight'][key[9:]] = list(set(hit['highlight'][key]))
-            yield item
-
-
-def search_result_actions(request, doc_types, es_results, position=None):
-    actions = {}
-    aggregations = es_results['aggregations']
-
-    return actions
-
-
-def format_facets(es_results, facets, used_filters, schemas, total, principals):
-    result = []
-    # Loading facets in to the results
-    if 'aggregations' not in es_results:
-        return result
-
-    aggregations = es_results['aggregations']
-    used_facets = set()
-    exists_facets = set()
-    for field, options in facets:
-        used_facets.add(field)
-        agg_name = field.replace('.', '-')
-        if agg_name not in aggregations:
-            continue
-        terms = aggregations[agg_name][agg_name]['buckets']
-        if len(terms) < 2:
-            continue
-        # internal_status exception. Only display for admin users
-        if field == 'internal_status' and 'group.admin' not in principals:
-            continue
-        facet_type = options.get('type', 'terms')
-        if facet_type == 'exists':
-            terms = [
-                {'key': 'yes', 'doc_count': terms['yes']['doc_count']},
-                {'key': 'no', 'doc_count': terms['no']['doc_count']},
-            ]
-            exists_facets.add(field)
-        result.append({
-            'type': facet_type,
-            'field': field,
-            'title': options.get('title', field),
-            'terms': terms,
-            'total': aggregations[agg_name]['doc_count']
-        })
-
-    # Show any filters that aren't facets as a fake facet with one entry,
-    # so that the filter can be viewed and removed
-    for field, values in used_filters.items():
-        if field not in used_facets and field.rstrip('!') not in exists_facets:
-            title = field
-            for schema in schemas:
-                if field in schema['properties']:
-                    title = schema['properties'][field].get('title', field)
-                    break
-            result.append({
-                'field': field,
-                'title': title,
-                'terms': [{'key': v} for v in values],
-                'total': total,
-                })
-
-    return result
-
-
-def normalize_query(request):
-    types = request.registry[TYPES]
-    fixed_types = (
-        (k, types[v].name if k == 'type' and v in types else v)
-        for k, v in request.params.items()
-    )
-    qs = urlencode([
-        (k.encode('utf-8'), v.encode('utf-8'))
-        for k, v in fixed_types
-    ])
-    return '?' + qs if qs else ''
-
-
-def iter_long_json(name, iterable, other):
-    import json
-
-    start = None
-
-    # Note: by yielding @graph (iterable) first, then the contents of result (other) *may* be altered based upon @graph
-    it = iter(iterable)
-    try:
-        first = next(it)
-    except StopIteration:
-        pass
-    else:
-        #yield json.dumps(first)
-        start = '{' + json.dumps(name) + ':['
-        yield start + json.dumps(first)
-        for value in it:
-            yield ',' + json.dumps(value)
-
-    if start is None: # Nothing has bee yielded yet
-        yield json.dumps(other)
-    else:
-        other_stuff = (',' + json.dumps(other)[1:-1]) if other else ''
-        yield ']' + other_stuff + '}'
 
 @view_config(route_name='search', request_method='GET', permission='search')
 def search(context, request, search_type=None, return_generator=False):
@@ -726,6 +196,9 @@ def search(context, request, search_type=None, return_generator=False):
             facets.append(audit_facet)
 
     query['aggs'] = set_facets(facets, used_filters, principals, doc_types)
+
+    # Search caching uses JSON as cache key, hence sorting the query
+    query = sort_query(query)
 
     # Decide whether to use scan for results.
     do_scan = size is None or size > 1000
