+++ src/snowflakes/tests/test_search.py
@@ -42,7 +42,7 @@ class FakeRequest(object):
 
 
 def test_set_filters():
-    from snowflakes.search import set_filters
+    from snovault.helpers.helper import set_filters
 
     request = FakeRequest((
         ('field1', 'value1'),
@@ -91,7 +91,7 @@ def test_set_filters():
 
 
 def test_set_filters_searchTerm():
-    from snowflakes.search import set_filters
+    from snovault.helpers.helper import set_filters
 
     request = FakeRequest((
         ('searchTerm', 'value1'),
@@ -138,7 +138,7 @@ def test_set_filters_searchTerm():
     'type', 'limit', 'mode',
     'format', 'frame', 'datastore', 'field', 'sort', 'from', 'referrer'])
 def test_set_filters_reserved_params(param):
-    from snowflakes.search import set_filters
+    from snovault.helpers.helper import set_filters
 
     request = FakeRequest((
         (param, 'foo'),
@@ -175,7 +175,7 @@ def test_set_filters_reserved_params(param):
 
 
 def test_set_filters_multivalued():
-    from snowflakes.search import set_filters
+    from snovault.helpers.helper import set_filters
 
     request = FakeRequest((
         ('field1', 'value1'),
@@ -230,7 +230,7 @@ def test_set_filters_multivalued():
 
 
 def test_set_filters_negated():
-    from snowflakes.search import set_filters
+    from snovault.helpers.helper import set_filters
 
     request = FakeRequest((
         ('field1!', 'value1'),
@@ -279,7 +279,7 @@ def test_set_filters_negated():
 
 
 def test_set_filters_audit():
-    from snowflakes.search import set_filters
+    from snovault.helpers.helper import set_filters
 
     request = FakeRequest((
         ('audit.foo', 'value1'),
@@ -329,7 +329,7 @@ def test_set_filters_audit():
 
 
 def test_set_filters_exists_missing():
-    from snowflakes.search import set_filters
+    from snovault.helpers.helper import set_filters
 
     request = FakeRequest((
         ('field1', '*'),
@@ -397,7 +397,7 @@ def test_set_filters_exists_missing():
 
 def test_set_facets():
     from collections import OrderedDict
-    from snowflakes.search import set_facets
+    from snovault.helpers.helper import set_facets
     facets = [
         ('type', {'title': 'Type'}),
         ('audit.foo', {'title': 'Audit'}),
@@ -483,7 +483,7 @@ def test_set_facets():
 
 def test_set_facets_negated_filter():
     from collections import OrderedDict
-    from snowflakes.search import set_facets
+    from snovault.helpers.helper import set_facets
     facets = [
         ('facet1', {'title': 'Facet 1'}),
     ]
@@ -523,7 +523,7 @@ def test_set_facets_negated_filter():
 
 def test_set_facets_type_exists():
     from collections import OrderedDict
-    from snowflakes.search import set_facets
+    from snovault.helpers.helper import set_facets
     facets = [
         ('field1', {'title': 'Facet 1', 'type': 'exists'}),
         ('field2', {'title': 'Facet 2', 'type': 'exists'}),
@@ -610,7 +610,7 @@ def test_set_facets_type_exists():
 
 
 def test_format_facets():
-    from snowflakes.search import format_facets
+    from snovault.helpers.helper import format_facets
     es_result = {
         'aggregations': {
             'field1': {
@@ -659,13 +659,13 @@ def test_format_facets():
 
 
 def test_format_facets_no_aggregations():
-    from snowflakes.search import format_facets
+    from snovault.helpers.helper import format_facets
     result = format_facets({}, [], [], [], 0, [])
     assert result == []
 
 
 def test_format_facets_skips_single_bucket_facets():
-    from snowflakes.search import format_facets
+    from snovault.helpers.helper import format_facets
     es_result = {
         'aggregations': {
             'field1': {
@@ -695,7 +695,7 @@ def test_format_facets_skips_single_bucket_facets():
 
 
 def test_format_facets_adds_pseudo_facet_for_extra_filters():
-    from snowflakes.search import format_facets
+    from snovault.helpers.helper import format_facets
     es_result = {
         'aggregations': {},
     }
